fix calendar token migration issue

+++ database/migrations/2025_07_29_144147_remove_calendar_token_from_users_table.php
@@ -11,9 +11,11 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('calendar_token');
-        });
+        if (Schema::hasColumn('users', 'calendar_token')) {
+            Schema::table('users', function (Blueprint $table) {
+                $table->dropColumn('calendar_token');
+            });
+        }
     }
 
     /**
